refactor: delete switch-system machinery — production path is straight-line only
The production driver (build_s_aff_kernels) and its four straight-line steps
no longer need the gate System, the streaming Pipeline, the worklist
garbler/evaluator, the journal, the compiled-garble arena, or the System
reference circuits. Deleted:

  system.rs, pipeline.rs, exec.rs, types.rs,
  comp_gc/{arena,convert,evaluator,garbler,ohe,program}.rs,
  build_s_aff_streaming, and all switch-system / streaming / primitive tests.

Relocated the three live helpers the kernels borrowed from deleted modules
into comp_gc/extract.rs (the NEON cast-pad unpackers + their differential
test, compute_sub_widths) and affine.rs (sample_cf_mask). Kept exactly the
a·x+b workload: label, crypto, hash, crt, comp_gc::{extract,fold}, it_gc,
affine, bitdecomp, and the benches/oracle tests.

13052 -> ~5900 lines; 105 tests pass (release, debug, count-hashes);
reference-workload instruction count and per-stage split unchanged.
